# optimize ppsnark (Nova Forward Port)