Ports https://github.com/microsoft/Nova/pull/252

<details>
<summary> Comparative CPU benchmark, sp. re: CompressedSNARK-Commitments</summary>

```
group                                                        dev                                    microsoft_252
-----                                                        ---                                    -------------
CompressedSNARK-Commitments-StepCircuitSize-0/Prove          1.01       6.4±0.07s            1.00       6.3±0.09s
CompressedSNARK-Commitments-StepCircuitSize-0/Verify         1.02    185.8±5.97ms            1.00    181.7±7.51ms
CompressedSNARK-Commitments-StepCircuitSize-121253/Prove     1.03      17.1±0.36s            1.00      16.6±0.49s
CompressedSNARK-Commitments-StepCircuitSize-121253/Verify    1.00   519.7±20.34ms            1.01   527.3±23.04ms
CompressedSNARK-Commitments-StepCircuitSize-22949/Prove      1.02       9.8±0.20s            1.00       9.6±0.17s
CompressedSNARK-Commitments-StepCircuitSize-22949/Verify     1.05   310.3±17.27ms            1.00   294.5±10.66ms
CompressedSNARK-Commitments-StepCircuitSize-252325/Prove     1.01      30.6±0.64s            1.00      30.3±0.76s
CompressedSNARK-Commitments-StepCircuitSize-252325/Verify    1.02   858.5±54.90ms            1.00   841.1±36.20ms
CompressedSNARK-Commitments-StepCircuitSize-55717/Prove      1.01      16.3±0.40s            1.00      16.1±0.42s
CompressedSNARK-Commitments-StepCircuitSize-55717/Verify     1.01   535.0±13.78ms            1.00   530.0±25.67ms
CompressedSNARK-Commitments-StepCircuitSize-6565/Prove       1.01       9.4±0.14s            1.00       9.3±0.10s
CompressedSNARK-Commitments-StepCircuitSize-6565/Verify      1.00    290.6±9.69ms            1.00   290.5±13.04ms
CompressedSNARK-StepCircuitSize-0/Prove                      1.00   765.2±10.17ms            1.01   772.5±10.60ms
CompressedSNARK-StepCircuitSize-0/Verify                     1.01     31.4±0.51ms            1.00     31.2±0.77ms
CompressedSNARK-StepCircuitSize-1038757/Prove                1.00      25.0±0.89s            1.00      25.0±0.72s
CompressedSNARK-StepCircuitSize-1038757/Verify               1.02   842.8±42.31ms            1.00   825.7±40.03ms
CompressedSNARK-StepCircuitSize-121253/Prove                 1.01       3.4±0.04s            1.00       3.3±0.09s
CompressedSNARK-StepCircuitSize-121253/Verify                1.00    128.7±4.59ms            1.00    128.1±7.67ms
CompressedSNARK-StepCircuitSize-22949/Prove                  1.01  1124.0±18.04ms            1.00  1113.5±16.20ms
CompressedSNARK-StepCircuitSize-22949/Verify                 1.03     49.0±1.34ms            1.00     47.5±2.11ms
CompressedSNARK-StepCircuitSize-252325/Prove                 1.01       6.3±0.23s            1.00       6.3±0.12s
CompressedSNARK-StepCircuitSize-252325/Verify                1.00   244.0±10.83ms            1.03   250.1±30.10ms
CompressedSNARK-StepCircuitSize-514469/Prove                 1.00      12.3±0.37s            1.01      12.4±0.27s
CompressedSNARK-StepCircuitSize-514469/Verify                1.00   498.8±18.11ms            1.00   500.5±10.38ms
CompressedSNARK-StepCircuitSize-55717/Prove                  1.01  1872.7±35.47ms            1.00  1855.5±18.05ms
CompressedSNARK-StepCircuitSize-55717/Verify                 1.03     72.9±4.26ms            1.00     70.7±2.82ms
CompressedSNARK-StepCircuitSize-6565/Prove                   1.00   764.8±11.47ms            1.00    762.7±9.43ms
CompressedSNARK-StepCircuitSize-6565/Verify                  1.00     31.3±0.77ms            1.01     31.6±0.62ms
```
</details>